fix(ui): prevent streaming reasoning reflow
